Change the replication status monitor app to use the HzMembershipMonitor class.  The App class instantiates it, and the EventMessageProcessor now supports messages of type 'MEMBERS' and 'MEMBER'.  Rename the ReplicationStatusEntryEvent to StatusEntryEvent so it's a bit more generic.
Changed the build to use the maven-shade-plugin to create an executable jar (as opposed to the maven-assembly-plugin) at Skye and Robert's suggestion.  Works well.

This code remains pretty hard-coded, but is functional.  We may want to rename the module to not be tied to replication if we make it more generic.