Do a bit more layout work on the Data View:
1) Make totals for DataONE prominent
2) Add proposed filters for the table
3) Add in a graph subview at the bottom

The intention is that the graph would change based on the sorting option in the table.  If the Data Files column were chosen, the Data Files graph would display.  Use colors to visually associate subviews.  If the table is filtered to a single MN, the graph changes to a time series for that MN, for the given column, defaulting to data files available.  Needs discussion. I'll get feedback from Amber.