Add a script that:
1) Dumps system metadata related tables from the Coordinating Node that the script is run on into a .sql dump file.
2) Modifies the SQL dump to append a prefix to entity names based on the hostname of the originating CN (e.g. ucsb_)
3) Transfers the SQL dump in bulk via rsync to the target machine where the dump will be restored to create the new per-CN tables where the tidy merging will happen.
Note that (4) is a manual step to run the SQL dump script on the target machine.
Also, this script requires that the postgres user on each machine the script runs on has an authorized_keys file in ~/.ssh that includes the public SSH key of the postgres user on the target machine, and the target machine postgres user has the private SSH key installed correctly.