- Added instructions for how to build and install the DataONE Test Utilities. - Rewrote the documentation for the stress tests. Factored out the information   that is common for all tests. Added documentation for the combined stress   tests. - Fixed formatting bug and wrong information in page describing how to create   test certificates. - Misc refactoring of the Test Utilities documentation. - Bumped version to 1.3.0. - Cleaned up script that generates test certificates. - Cleaned up the way the stress tests handle the lists of public and private objects, and the lists of subjects with access to objects. - Cleaned up the stress tests, factoring things out to the common transaction class. - Added readme in the main folder for the stress tests pointing to the documentation. - Updated the stress tests to match the change in common, where I factored checksum function out to their own module. - Wrote a new program called the Replication Tester and documentation for it.
Overview of the Replication Tester:

The Replication Tester can test replication processing on both source and
destination MNs. The tester takes on the role of a CN during testing. In
addition, when testing a source MN, it takes on the role of a destination MN and
vice versa.


The tester has an embedded web server that runs asynchronously, so that it can
accept and check the calls that the MN being tested would normally issue to the
CN and other MN.

The MN being tested must be configured to accept the certificates issued by the
tester as if they come from real CNs and MNs. The MN being tested must be
configured to call back to the Replication Tester instead of to a CN or MN.


- Bumped versions of dependencies.